Improvements to the iOS Phone/Tasks implementations.
The ContentIdHelper converts the app name to a valid name for the appstore link by removing all non-alphanumeric characters. This technique was described in an answer to this question: http://stackoverflow.com/questions/433907/how-to-link-to-apps-on-the-app-store

I've tested it with http appstore links and it appears to work in all cases, once I worked out that the '&' character needs to be output as "and"!

*Examples*
"He-Man: The Most Powerful Game in the Universe(tm)" - http://itunes.com/app/HeManTheMostPowerfulGameintheUniverse
"Millie's Book of Tricks and Treats, Volume 2" - http://itunes.com/app/MilliesBookofTricksandTreatsVolume2
"Living Earth - Clock & Weather" - http://itunes.com/app/LivingEarthClockandWeather